Strategy pattern, Mappers

+++ src/main/java/com/example/demo/Game.java
@@ -11,13 +11,19 @@ import javax.persistence.*;
 @Setter
 @Getter
 public class Game {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int bet;
-    private int prize;
+
+    private float bet;
+
+    private float prize;
+
     private Boolean isFreeRoundWon;
+
     private Boolean isFree; // ?
+
     @ManyToOne
     @JoinColumn(name = "player_id")
     private Player player;
