+++ src/main/java/com/example/demo/dto/PlayerDto.java
@@ -8,8 +8,12 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class PlayerDto {
+
     private Long playerId;
+
     private String name;
-    private int balance;
+
+    private float balance;
+
     private boolean isNextFree;
 }
